Feat: fixed image upload + guild main page
- added guild main page to make redirecting simpler.
- image upload using multer fixed.

+++ server/src/controllers/guildController.ts
@@ -79,8 +79,8 @@ const getGuild = async (req: Request, res: Response) => {
 };
 
 const createGuild = async (req: Request, res: Response) => {
-  // const userId = req.auth.userId;
-  const userId = "user_2kvgB9d6HPZNSZGsGDf02nYSx12";
+  const userId = req.auth.userId;
+  // const userId = "user_2kvgB9d6HPZNSZGsGDf02nYSx12";
   const { name } = req.body as { name: string };
 
   if (!name) {
@@ -144,6 +144,22 @@ const leaveGuild = async (req: Request, res: Response) => {
   const { guildId } = req.params as { guildId: string };
 
   try {
+    const isOwner = await prisma.member.findFirst({
+      where: {
+        userId: userId,
+        guildId: guildId,
+        role: "OWNER",
+      },
+    });
+
+    if (isOwner) {
+      return res
+        .status(403)
+        .json({
+          error: "You cannot leave a guild you own. You'd have to delete it.",
+        });
+    }
+
     await prisma.member.deleteMany({
       where: {
         userId: userId,
@@ -195,4 +211,95 @@ const deleteGuild = async (req: Request, res: Response) => {
   }
 };
 
-export { getAllGuilds, createGuild, getGuild, leaveGuild, deleteGuild };
+const updateGuild = async (req: Request, res: Response) => {
+  const userId = req.auth.userId;
+  // const userId = "user_2kvgB9d6HPZNSZGsGDf02nYSx12";
+  const { guildId } = req.body as { guildId: string };
+  const { name } = req.body as { name: string };
+
+  console.log(req.body);
+
+  const isOwnerAdmin = await prisma.member.findFirst({
+    where: {
+      userId: userId,
+      guildId: guildId,
+      OR: [
+        {
+          role: "OWNER",
+        },
+        {
+          role: "ADMIN",
+        },
+      ],
+    },
+  });
+
+  console.log("test");
+
+  console.log(isOwnerAdmin);
+
+  if (!isOwnerAdmin) {
+    return res
+      .status(403)
+      .json({ error: "You are not the owner or an admin of this guild" });
+  }
+
+  if (!name) {
+    return res.status(400).json({ error: "Guild name is required" });
+  }
+
+  const file = req.file as Express.MulterS3.File;
+
+  console.log(file);
+
+  //if there is an image uploaded
+  if (file.originalname !== "undefined") {
+    if (file.mimetype.split("/")[0] !== "image") {
+      return res
+        .status(400)
+        .json({ error: "Invalid image type. Valid types are: png, jpg, jpeg" });
+    }
+
+    try {
+      const updatedGuild = await prisma.guild.update({
+        where: {
+          id: guildId,
+        },
+        data: {
+          name: name,
+          imageUrl: file.location,
+        },
+      });
+
+      res.json(updatedGuild);
+    } catch (error) {
+      console.error("Error updating guild:", error);
+      res.status(500).json({ error: "Failed to update guild" });
+    }
+  } else {
+    try {
+      const updatedGuild = await prisma.guild.update({
+        where: {
+          id: guildId,
+        },
+        data: {
+          name: name,
+        },
+      });
+
+      res.json(updatedGuild);
+    } catch (error) {
+      console.error("Error updating guild:", error);
+      res.status(500).json({ error: "Failed to update guild" });
+    }
+  }
+};
+
+export {
+  getAllGuilds,
+  createGuild,
+  getGuild,
+  leaveGuild,
+  deleteGuild,
+  updateGuild,
+};
